% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.support-fr.R
\name{lang.support.fr}
\alias{lang.support.fr}
\title{Language support for French}
\usage{
lang.support.fr()
}
\description{
This function adds support for French to the koRpus package. You should not
need to call it manually, as that is done automatically when this package is
loaded.
}
\details{
In particular, this function adds the following:
\itemize{
 \item \code{lang}: The additional language "fr" to be used with koRpus
 \item \code{treetag}: The additional preset "fr",
      implemented according to the respective TreeTagger[1] script
 \item \code{POS tags}: An additional set of tags,
      implemented using the documentation for the corresponding
   TreeTagger parameter set[2]
}

To also be able to use hyphenation, the package sylly.fr needs to exist and be loaded.
}
\examples{
\dontrun{
lang.support.fr()
}
}
\references{
[1] \url{http://www.cis.uni-muenchen.de/~schmid/tools/TreeTagger/}

[2] \url{http://www.cis.uni-muenchen.de/~schmid/tools/TreeTagger/data/french-tagset.html}
}
